Refactor BoardStateSwitcher and replace FloatingPanel with PanelOverlay
- Removed unused state management and dialog components from BoardStateSwitcher, simplifying its functionality.
- Introduced a new method for adding custom states with automatic naming, enhancing user experience.
- Replaced the deprecated FloatingPanel with a new PanelOverlay component for improved dialog management.
- Updated the AI chat panel to use VMAiChatPanel, streamlining the chat interface and maintaining consistency across panels.

These changes enhance the overall structure and usability of the editor, providing a more cohesive and responsive interface.
